refact(lib): move lib code files
