+++ lib/phonetic-symbol.js
