+++ lib/pinyin.js
