+++ lib/web-pinyin.js
@@ -21,8 +21,8 @@ function buildPinyinCache(dict_combo){
   return uncomboed;
 }
 
-const PINYIN_DICT = buildPinyinCache(require("./data/dict-zi-web"));
-const Pinyin = require("./src/pinyin");
+const PINYIN_DICT = buildPinyinCache(require("../data/dict-zi-web"));
+const Pinyin = require("./pinyin");
 const pinyin = new Pinyin(PINYIN_DICT);
 
 module.exports = pinyin.convert.bind(pinyin);
